backend/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
dlib==19.24.2
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==21.2.0
mysqlclient==2.2.0
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
pypng==0.20220715.0
pytz==2023.3.post1
qrcode==7.4.2
reportlab==4.0.6
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
